Build system: switch to CMake
I am going to be using SDL3 for the GUI, so handling dependencies like
libraries is going to be much more comfortable with CMake.

+++ src/utils.c
@@ -1,8 +1,8 @@
 #include <stdio.h>
 #include <stdlib.h>
 
-#include "utils.h"
-#include "cli.h"
+#include "xm65/utils.h"
+#include "xm65/cli.h"
 
 char* XM65_ByteToBin(uint8_t byte) {
     static char bits[9];  // 8 bits + null terminator
@@ -33,13 +33,13 @@ void XM65_OpenFile(const char *filename, char **buffer, long *filesize) {
     *filesize = ftell(file);
     rewind(file);
 
-    *buffer = (char *) malloc(*filesize);
+    *buffer = (char *) malloc((size_t) *filesize);
     if (!*buffer) {
         fclose(file);
         XM65_ThrowError(XM65_ERROR_MEM_ALLOC);
     }
 
-    size_t read_bytes = fread(*buffer, 1, *filesize, file);
+    size_t read_bytes = fread(*buffer, 1, (size_t) *filesize, file);
     if ((long) read_bytes != *filesize) {
         free(*buffer);
         fclose(file);
